+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Develop
 
-- Added `LWBTN_CFG_GET_STATE_MODE` to control *get state* mode
+- Add `LWBTN_CFG_GET_STATE_MODE` to control *get state* mode
+- Add option to check if button is currently active (after debounce period has elapsed) 
 
 ## v0.0.1